requirements.txt
Flask >= 3.0, <= 3.1
brother_ql >= 0.9, <= 1.0
Pillow == 10.*
pylibdmtx == 0.1.*
python-dotenv == 1.*
gunicorn